Cambio de estructura

+++ src/main/java/com/example/wayra/repository/CiudadRepository.java
@@ -3,5 +3,5 @@ package com.example.wayra.repository;
 import com.example.wayra.model.Ciudad;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CiudadRepository extends JpaRepository<Ciudad, Long> {
+public interface CiudadRepository extends JpaRepository<Ciudad, Integer> {
 }
