+++ src/main/java/com/example/wayra/repository/ClienteRepository.java
@@ -1,7 +1,8 @@
 package com.example.wayra.repository;
 
 import com.example.wayra.model.Cliente;
+import com.example.wayra.model.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClienteRepository extends JpaRepository<Cliente, Long> {
+public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
 }
